% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFW_waterShp.R
\name{DFW_waterShp}
\alias{DFW_waterShp}
\title{Dallas-Fort Worth Metropolitan Area Water Area}
\format{
A SpatialPolygonsDataFrame with block group observations and
1 variable:
\describe{
\item{fclass}{Class of water feature, \code{reservoir} or \code{riverbank}}
}
}
\description{
A spatial data frame containing the resevoir and river features in NCTCOG counties.
}
