% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapColorQual.R
\name{mapColorQual}
\alias{mapColorQual}
\title{Function: Maps a qualitative theme for a maximum of 12 categories}
\usage{
mapColorQual(var.name, shape, map.title="", legend.title=deparse(substitute(var.name)),
                    legend.pos="bottomleft", legend.cex=1, add.to.map=FALSE)
}
\arguments{
\item{var.name}{A factor, perhaps with NA's, to be mapped with a maximum of 12 categories.
If the factor is in a data-frame, then the data-frame must be explicitly
referred to, e.g., \code{df$var}}

\item{shape}{An existing spatial polygon or spatial polygon data-frame}

\item{map.title}{Character string with map title}

\item{legend.title}{Character string with legend title (default=\code{var.name})}

\item{legend.pos}{Location of legend in the map frame (default=\code{"bottomleft"})}

\item{legend.cex}{Relative font size of the legend}

\item{add.to.map}{Logical to start a new map frame if \code{FALSE} or overlay onto an
existing map frame if \code{TRUE}}
}
\value{
\code{NULL}
}
\description{
\code{mapColorQual} generates a map of a qualitative variable
}
\details{
The function \code{mapColorQual} maps a \emph{categorical variable} with a
  \emph{set of distinct colors}. A legend is generated. The maximum number of
  valid factor levels should not exceed 12. NA's are permitted.
}
\examples{
validTractShp <- DFW_tractShp[DFW_tractShp$COUNTY=="Dallas County",]
mapColorQual(validTractShp$CongLevel, validTractShp,
             map.title="Congestion in Dallas County",
             legend.title="Congestion")
}
\author{
Michael Tiefelsdorf <tiefelsdorf@utdallas.edu>
}
