% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_hist.R
\name{palette_hist}
\alias{palette_hist}
\title{Function: creates a histogram of a color palette with set breaks}
\usage{
palette_hist(class.intervals, pal=NULL, title = "")
}
\arguments{
\item{class.intervals}{\code{ClassIntervals} object or named list with items
\code{class.intervals} and \code{pal}}

\item{pal}{a color palette from \code{RColorBrewer:brewer.pal}}

\item{title}{optional title for the histogram}

\item{equal.width}{Logical, determines if equal width bins should be used
(default=\code{FALSE})}

\item{bins}{numeric number of bins to use (forces \code{equal.width=T})}
}
\value{
\code{NULL}
}
\description{
\code{palette_hist} generates a histogram of a color palette
of an interval scaled variable
}
\details{
The function \code{palette_hist} maps an \emph{interval scaled
variable} by a \emph{sequetial color ramp}.
}
\examples{
ramp <- mapColorRamp(tractShp, 'bad1500D', breaks=9,
             map.title="Density of Convenience Stores in Dallas County\nbw=1500 meters",
             legend.title="Junk Food", output.breaks=TRUE)

palette_hist(ramp)

palette_hist(ramp$class.intervals, ramp$pal)

}
\author{
Mae Hutchison <mah230002@utdallas.edu>
}
