% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewPaletteSpread.R
\name{viewPaletteSpread}
\alias{viewPaletteSpread}
\title{Function: creates a histogram of a color palette with set breaks}
\usage{
viewPaletteSpread(class.intervals, pal=NULL, title="",
equal.width=TRUE, bins=NULL, density=TRUE, density.bw="nrd",
density.kernel="gaussian", density.adjust=1, ColorRamp.breaks=FALSE)
}
\arguments{
\item{class.intervals}{\code{ClassIntervals} object or named list with items
\code{class.intervals} and \code{pal}}

\item{pal}{a color palette from \code{RColorBrewer:brewer.pal}}

\item{title}{optional title for the histogram}

\item{equal.width}{Logical, determines if equal width bins should be used
(default=\code{TRUE})}

\item{bins}{numeric number of bins to use (forces \code{equal.width=T})}

\item{density}{Logical, displays estimated density curve over plot (default=\code{TRUE})}

\item{density.bw}{bandwidth argument passed to \code{stats::density()} function}

\item{density.kernel}{kernel argument passed to \code{stats::density()} function}

\item{density.adjust}{bandwidth adjustment multiplier passed to \code{stats::density()} function}

\item{ColorRamp.breaks}{Logical, displays exact breakpoints that the color ramp changes}
}
\value{
\code{NULL}
}
\description{
\code{viewPaletteSpread} generates a histogram of a color palette
of an interval scaled variable
}
\details{
The function \code{viewPaletteSpread} maps an \emph{interval scaled
variable} by a \emph{sequetial color ramp}.
}
\examples{
ramp <- mapColorRamp(DFW_tractShp, 'PERCAPINC', breaks=9,
             map.title="Per Capita Income",
             legend.title="Income", output.breaks=TRUE)

viewPaletteSpread(ramp)

viewPaletteSpread(ramp$class.intervals, ramp$pal)

}
\author{
Mae Hutchison <mah230002@utdallas.edu>
}
