% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bndShp.R
\docType{data}
\name{bndShp}
\alias{bndShp}
\title{Boundary of Dallas County, TX}
\description{
One polygon in the longitude and latitude format (see
  \code{proj4string=CRS("+proj=longlat +ellps=WGS84")}).
}
\examples{
library(maptools)
validTractShp <- tractShp[!is.na(tractShp$BUYPOW), ]  # Remove 2 tracts with NA's
plot(tractShp, col="white", border="white", axes=TRUE,
     main="Dallas Census Tracts with Food Deserts")
plot(validTractShp, col="ivory2", border="white", add=TRUE)
plot(lakesShp, col="skyblue", border="skyblue",add=TRUE)
plot(hwyShp, col="cornsilk3", lwd=3, add=TRUE)
plot(foodDesertShp, border="magenta",lwd=2, add=TRUE)
plot(bndShp, border="black", add=TRUE)
box()

}
